[ALPS08543716] audio: fix 32bit project build error
fix 32bit project build error

MTK-Commit-Id: 6f59894558b4ca83233002743ad25cc8dea81e8f

CR-Id: ALPS08543716
Feature: alsa_sound_soc

+++ sound/soc/mediatek/common/mtk-afe-platform-driver.h
@@ -32,6 +32,6 @@ void mtk_afe_pcm_free(struct snd_soc_component *component, struct snd_pcm *pcm);
 
 int mtk_afe_combine_sub_dai(struct mtk_base_afe *afe);
 int mtk_afe_add_sub_dai_control(struct snd_soc_component *component);
-unsigned long word_size_align(unsigned long in_size);
+unsigned long long word_size_align(unsigned long long in_size);
 #endif
